Fixed no overlay images saved error
There was an issue that the first time you ran the script, no annotated images with the watershed overlay would be saved, but the second time you ran it, it would. This was because the directory wasn't created until after the image was processed, so the first time around, there was no directory in which to save the images.